+++ src/main/java/net/frozenblock/wilderwild/WilderWild.java
@@ -59,7 +59,7 @@ import net.frozenblock.wilderwild.registry.RegisterSensorTypes;
 import net.frozenblock.wilderwild.registry.RegisterSounds;
 import net.frozenblock.wilderwild.registry.RegisterWorldgen;
 import net.frozenblock.wilderwild.registry.WilderRegistry;
-import net.frozenblock.wilderwild.world.generation.WilderWorldGen;
+import net.frozenblock.wilderwild.world.modification.WilderWorldGen;
 import org.jetbrains.annotations.NotNull;
 
 public final class WilderWild extends FrozenModInitializer implements FrozenMobCategoryEntrypoint {