+++ src/main/java/net/frozenblock/wilderwild/world/modification/WilderMusic.java
@@ -16,7 +16,7 @@
  * along with this program; if not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.frozenblock.wilderwild.world.generation;
+package net.frozenblock.wilderwild.world.modification;
 
 import net.fabricmc.fabric.api.biome.v1.BiomeModifications;
 import net.fabricmc.fabric.api.biome.v1.ModificationPhase;